Some IOP fixes; now loops waiting for an SIF value to change
Small changes, but big steps! We're getting closer and closer to fully LLEing the BIOS.

Current commit gives the IOP the value it wants, which sends it to an infinite loop; this is surely intentional as the IOP won't be doing any work after initialization unless the EE or something else sends an interrupt.