Added "skipmpeg" hack
This allows us to get rid of the previous hack for Atelier Iris, which is somewhat good as this one is generic.

Why have this hack? FMVs are only a secondary consideration compared to getting the rest of the emulator functional. Once the core emulation is good, we can work on perfecting movie playback.