Add screenshot functionality (#179)
Add the ability to take screenshots 2 ways:
- By pressing `f8`
- `Window -> Take Screenshot`

The directory in which screenshots are saved can be changed by the user `Options -> Settings -> Paths -> Screenshots`. When it's not set, it's the user's home directory.

The format of the screenshot filename is as follows:
`yyyyMMddHHmmsszzz.png`.

Currently uses default settings for compression.
PNG used because JPEG is garbage.