gs-int: specify color format for z write when color write is z format and enabled
Confirmed by hw tests performed by water111.

The GS divides the block ordering of the z and color writes by format. This allows both writes to happen at the same time.
Color writes are performed on the left side of the page while z writes are performed on the right side.

However, in the event that a z format for frame, the gs will swap the zbuf format for a color one.

Fix power drome.