CDVD: Changes to N status and drive status
It was originally hypothesized that bit 6 of N_status represented "ready to receive commands." However, Jimmy Neutron: Boy Genius crashes with the old setup. It sends sceCdRead twice to CDVDMAN to read two separate files: a level map and a music file. When the music file load was occurring, CDVDMAN would return an error on the second sceCdRead as bit 6 was cleared. The game would think that this file had loaded successfully, causing a crash due to reading from uninitialized memory.

This commit changes this bit to always be on and adjusts drive status to be set to SEEKING when a seek starts. When the N status check passes, CDVDMAN will loop waiting for drive status to become paused, rather than returning an error. Thus, the second sceCdRead successfully completes eventually, allowing the game to work.