Scheduler: Remove dependency on Emulator
This allows for scheduler functions to be defined anywhere, rather than being restricted to the Emulator class.

This commit also adds a "register_function" method to Scheduler. This makes serialization of events easier as we no longer need to hardcode event IDs.

Finally, events now have an optional 64-bit parameter.