Set bit 7 of spu status reg on DMA pause
This seems more like a DMA_ready bit.

When asked to stop a transfer libsd turns off the busy bit on the dma control
register which pauses the DMA transfer. It then proceeds to loop waiting
for bit 7 of spu status to turn to 1.

This is what's happening to Gregory Horror Show.